refactor: add styles to 404 page
